Add DateTimeHelper for culture-specific date formatting and refactor ProductInfo
